+++ CHANGELOG.md
@@ -11,6 +11,7 @@ around 150k
 - merged mimic_base and mimic, and added a seperate mimic_common trait.  Now the Schema types that are shared between
 mimic and mimic_design are properly wrapped and kept DRY
 - redid the crate layout so that people only have to include mimic
+- removed rust-decimal, now we use a native FixedE8 type that should handle all the cases
 
 ## [0.5.3]
 - full support for WHERE logic trees, with simplification and DeMorgan's laws (thanks GCSE Computing!)